Also add back #3171 by @shouyiwang