@Dampfinchen 2023-08-18
Can you please add --mul-mat-q parameter for llama.cpp python? The new mulmat kernels save a lot of vram and speed up prompt processing.