@jllllll 2023-08-18
Made a PR for llama-cpp-python-cuda to merge into the bump-llamacpp branch: https://github.com/oobabooga/text-generation-webui/pull/3614

Didn't notice this PR until I had already made one.