@oobabooga 2023-08-18
@Dampfinchen yes, done. I wasn't aware of this parameter and it does seem to make things somewhat faster:

* Before:  Output generated in 9.22 seconds (21.58 tokens/s, 199 tokens, context 1501, seed 1395691830)

* After: Output generated in 8.76 seconds (22.72 tokens/s, 199 tokens, context 1501, seed 259568941)

@jllllll thanks for the wheel update! No problem, I'll merge this PR and then yours.